% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_variables.R
\name{get_acs_vars}
\alias{get_acs_vars}
\title{Get all ACS variable codes needed for a set of racial groups}
\usage{
get_acs_vars(groups = c("white", "black", "hispanic", "aapi", "native"))
}
\arguments{
\item{groups}{Character vector of group names}
}
\value{
Character vector of ACS variable codes
}
\description{
Get all ACS variable codes needed for a set of racial groups
}
\keyword{internal}
