% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f5_plot.R
\name{scales_pct_label}
\alias{scales_pct_label}
\title{Simple percentage label function (avoids scales dependency)}
\usage{
scales_pct_label(x)
}
\description{
Simple percentage label function (avoids scales dependency)
}
\keyword{internal}
